requirements.txt
streamlit>=1.28.0
stable-baselines3>=2.0.0
gymnasium[atari]>=0.29.0
gym>=0.26.0
ale-py>=0.8.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
matplotlib>=3.7.0
plotly>=5.15.0
autorom[accept-rom-license]>=0.4.2
shimmy>=0.2.1